backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain_community
langchain-core
groq
python-dotenv
